add base test

+++ test/mapper/test.mapper.go
@@ -8,7 +8,7 @@ import (
 
 type TestMapper struct {
 	SelectInsert                     func(entity entity.TestEntity) (id int, err error)
-	SelectInsertPointer              func(entity *entity.TestEntityPointer) (id int8, err error)
+	SelectInsertPointer              func(entity *entity.TestEntityPointer) (id int32, err error)
 	SelectInsertForeachSlice         func(testEntity entity.TestEntity, enums []string) (id int16, err error)
 	SelectInsertForeachSlicePointer  func(testEntity *entity.TestEntityPointer, enums *[][]*string) (id int32, err error)
 	SelectInsertForeachMap           func(testEntity entity.TestEntity, enums map[string][]string) (id uint, err error)
